implemented hook

+++ src/client/client.hpp
@@ -29,7 +29,7 @@ class client {
 		netsi::client_network_manager _network_manager;
 		std::unique_ptr<renderer> _renderer;
 		std::shared_ptr<netsi::peer> _peer;
-		std::uint8_t _last_actions;
+		std::uint16_t _last_actions;
 		char _local_player_id;
 };
